Evaluation_of_RAG/Evaluation_Rag_Gurkirt_Kaur/requirements.txt
streamlit
pandas
numpy
sentence-transformers
faiss-cpu
groq
python-dotenv
PyJWT
spacy
transformers
torch
plotly
